Refactor modelzoo
