+++ lucid/modelzoo/caffe_models/README.md
@@ -1,0 +1,3 @@
+Models imported from caffe.
+
+See `lucid/scripts/import_caffe/` for details on how to import more.
