+++ lucid/modelzoo/caffe_models/__init__.py
@@ -1,0 +1,2 @@
+from lucid.modelzoo.caffe_models.InceptionV1 import *
+from lucid.modelzoo.caffe_models.others import *
