+++ lucid/modelzoo/other_models/__init__.py
@@ -1,0 +1,1 @@
+from lucid.modelzoo.other_models.InceptionV1 import InceptionV1
