+++ lucid/modelzoo/slim_models/__init__.py
@@ -1,0 +1,4 @@
+from lucid.modelzoo.slim_models.Inception import *
+from lucid.modelzoo.slim_models.ResnetV1 import *
+from lucid.modelzoo.slim_models.ResnetV2 import *
+from lucid.modelzoo.slim_models.others import *
